build: adopt @olivierzal/homey-kit 3.1.0
