+++ package.json
@@ -32,7 +32,7 @@
   },
   "prettier": "@olivierzal/configs/prettier",
   "dependencies": {
-    "@olivierzal/homey-kit": "2.2.0",
+    "@olivierzal/homey-kit": "3.1.0",
     "@olivierzal/melcloud-api": "47.0.0",
     "source-map-support": "^0.5.21",
     "temporal-polyfill": "^1.0.1"
