+++ tests/unit/api.test.ts
@@ -1,3 +1,4 @@
+import type { DriverSetting } from '@olivierzal/homey-kit/manifest'
 import type * as Classic from '@olivierzal/melcloud-api/classic'
 import type * as Home from '@olivierzal/melcloud-api/home'
 import type { Homey } from 'homey/lib/Homey'
@@ -11,7 +12,6 @@ import {
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
 import type { DeviceSettings, Settings } from '../../types/device-settings.mts'
-import type { DriverSetting } from '../../types/driver-settings.mts'
 import type {
   ErrorLogQueryParams,
   FormattedErrorLog,
